+++ src/beta-negative-binomial-distribution.cpp
@@ -1,7 +1,22 @@
 #include <Rcpp.h>
-#include "namespace.h"
 #include "const.h"
 #include "shared.h"
+
+using std::pow;
+using std::sqrt;
+using std::abs;
+using std::exp;
+using std::log;
+using std::floor;
+using std::ceil;
+using std::sin;
+using std::cos;
+using std::tan;
+using std::atan;
+using Rcpp::IntegerVector;
+using Rcpp::NumericVector;
+using Rcpp::NumericMatrix;
+
 
 /*
 *  Beta-negative binomial distribution
@@ -51,7 +66,7 @@ double cdf_bnbinom(double k, double r, double alpha, double beta) {
     return 1;
   double p_tmp = 0;
   for (int j = 0; j < k+1; j++)
-    p_tmp += exp(logpmf_bnbinom(j, r, alpha, beta))*P_NORM_CONST;
+    p_tmp += exp(logpmf_bnbinom(static_cast<double>(j), r, alpha, beta))*P_NORM_CONST;
   return p_tmp/P_NORM_CONST;
 }
 
@@ -118,17 +133,18 @@ NumericVector cpp_pbnbinom(
     }
     
     double mx = finite_max(x);
-    NumericVector p_tab(mx+1);
+    NumericVector p_tab(static_cast<int>(mx)+1);
     
-    p_tab[0] = exp(logpmf_bnbinom(0, size[0], alpha[0], beta[0]))*P_NORM_CONST;
+    p_tab[0] = exp(logpmf_bnbinom(0.0, size[0], alpha[0], beta[0]))*P_NORM_CONST;
     for (int j = 1; j < mx+1; j++)
-      p_tab[j] = p_tab[j-1] + exp(logpmf_bnbinom(j, size[0], alpha[0], beta[0]))*P_NORM_CONST;
+      p_tab[j] = p_tab[j-1] + exp(logpmf_bnbinom(static_cast<double>(j),
+                                                 size[0], alpha[0], beta[0]))*P_NORM_CONST;
     
     for (int i = 0; i < n; i++) {
       if (std::isinf(x[i])) {
         p[i] = 1;
-      } else if (x[i] == floor(x[i]) && x[i] >= 0) {
-        p[i] = p_tab[(int)x[i]]/P_NORM_CONST;
+      } else if (isInteger(x[i]) && x[i] >= 0) {
+        p[i] = p_tab[static_cast<int>(x[i])]/P_NORM_CONST;
       } else {
         p[i] = 0;
       }