+++ src/mixture-of-normal-distributions.cpp
@@ -1,7 +1,21 @@
 #include <Rcpp.h>
-#include "namespace.h"
 #include "const.h"
 #include "shared.h"
+
+using std::pow;
+using std::sqrt;
+using std::abs;
+using std::exp;
+using std::log;
+using std::floor;
+using std::ceil;
+using std::sin;
+using std::cos;
+using std::tan;
+using std::atan;
+using Rcpp::IntegerVector;
+using Rcpp::NumericVector;
+using Rcpp::NumericMatrix;
 
 
 // [[Rcpp::export]]
@@ -37,7 +51,7 @@ NumericVector cpp_dmixnorm(
       if (sigma(i % ns, j) <= 0)
         wrong_param = true;
     }
-    if (!tol_equal(alpha_tot/P_NORM_CONST, 1) || wrong_param) {
+    if (!tol_equal(alpha_tot/P_NORM_CONST, 1.0) || wrong_param) {
       Rcpp::warning("NaNs produced");
       p[i] = NAN;
     }
@@ -84,7 +98,7 @@ NumericVector cpp_pmixnorm(
       if (sigma(i % ns, j) <= 0)
         wrong_param = true;
     }
-    if (!tol_equal(alpha_tot/P_NORM_CONST, 1) || wrong_param) {
+    if (!tol_equal(alpha_tot/P_NORM_CONST, 1.0) || wrong_param) {
       Rcpp::warning("NaNs produced");
       p[i] = NAN;
     }
@@ -117,25 +131,37 @@ NumericVector cpp_rmixnorm(
   
   bool wrong_param, j_set;
   double u, p_tmp;
-  int jj;
   NumericVector prob(k);
   
   for (int i = 0; i < n; i++) {
+    int jj;
     wrong_param = false;
-    j_set = false;
     u = R::runif(0, 1)*P_NORM_CONST;
     p_tmp = P_NORM_CONST;
+    
     for (int j = k-1; j >= 0; j--) {
       p_tmp -= alpha(i % na, j)*P_NORM_CONST;
-      if (sigma(i % ns, j) <= 0)
+      if (sigma(i % ns, j) < 0 || alpha(i % na, j) < 0 || alpha(i % na, j) > 1) {
         wrong_param = true;
-      if (!j_set && u > p_tmp) {
+        break;
+      }
+      if (u > p_tmp) {
         jj = j;
-        j_set = true;
+        break;
       }
     }
     
-    if (!tol_equal(p_tmp/P_NORM_CONST, 0) || wrong_param) {
+    if (!wrong_param && jj > 0) {
+      for (int j = jj-1; j >= 0; j--) {
+        p_tmp -= alpha(i % na, j)*P_NORM_CONST;
+        if (sigma(i % ns, j) < 0 || alpha(i % na, j) < 0 || alpha(i % na, j) > 1) {
+          wrong_param = true;
+          break;
+        }
+      } 
+    }
+
+    if (wrong_param || !tol_equal(p_tmp/P_NORM_CONST, 0.0)) {
       Rcpp::warning("NaNs produced");
       x[i] = NAN;
     } else {