+++ src/slash-distribution.cpp
@@ -1,6 +1,20 @@
 #include <Rcpp.h>
-#include "namespace.h"
 #include "shared.h"
+
+using std::pow;
+using std::sqrt;
+using std::abs;
+using std::exp;
+using std::log;
+using std::floor;
+using std::ceil;
+using std::sin;
+using std::cos;
+using std::tan;
+using std::atan;
+using Rcpp::IntegerVector;
+using Rcpp::NumericVector;
+using Rcpp::NumericMatrix;
 
 
 /*
@@ -20,7 +34,7 @@ double pdf_slash(double x, double mu, double sigma) {
     return NAN;
   }
   double z = (x - mu)/sigma;
-  return ((phi(0) - phi(z))/pow(z, 2))/sigma;
+  return ((phi(0.0) - phi(z))/pow(z, 2.0))/sigma;
 }
 
 double cdf_slash(double x, double mu, double sigma) {
@@ -32,7 +46,7 @@ double cdf_slash(double x, double mu, double sigma) {
   if (z == 0)
     return 0.5;
   else
-    return Phi(z) - (phi(0) - phi(z))/z;
+    return Phi(z) - (phi(0.0) - phi(z))/z;
 }
 
 double rng_slash(double mu, double sigma) {